pyproject.toml
[project]
name = "aieng"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Marko Elez", email = "markoelez7@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "gitpython>=3.1.0",
    "tomli>=2.0.1",
    "tomli-w>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
aieng = "aieng:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.12.3",
    "pytest>=8.0.0",
    "ty>=0.0.2",
]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["."]
